feat: imrpoved tui pallete
